Many bug fixes, but not done
Unfortunately, my hack to try and keep the same schema of the cached values for both scores and ratings has backfired. While unlikely to occur in actuality, using the key format <user>.<cluster> means that checking of the user is in the key when the cluster is also of the same type (i.e., both are integers), then I run the risk of confusing the user and the cluster. Thus, I am going to refactor so that instead of different keys, I'll just add different getters for each type of dict schema. Once this is done, I think it should be all good!

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>